refactor: move Timeline component to the components folder
